restructured project, no changes otherwise
