--- converter/text.txt
@@ -1,39 +1,0 @@
-//"allow_cabin_overheat_protection": true,
-//"auto_seat_climate_left": false,
-//"auto_seat_climate_right": false,
-//"auto_steering_wheel_heat": false,
-//"battery_heater": false,
-//"battery_heater_no_power": null,
-//"bioweapon_mode": false,
-//"cabin_overheat_protection": "On",
-//"cabin_overheat_protection_actively_cooling": true,
-//"climate_keeper_mode": "off",
-//"cop_activation_temperature": "High",
-//"defrost_mode": 0,
-//"driver_temp_setting": 21,
-//"fan_status": 0,
-//"hvac_auto_request": "On",
-//"inside_temp": 38.4,
-//"is_auto_conditioning_on": true,
-//"is_climate_on": false,
-//"is_front_defroster_on": false,
-//"is_preconditioning": false,
-//"is_rear_defroster_on": false,
-//"left_temp_direction": -293,
-//"max_avail_temp": 28,
-//"min_avail_temp": 15,
-//"outside_temp": 36.5,
-//"passenger_temp_setting": 21,
-//"remote_heater_control_enabled": false,
-//"right_temp_direction": -276,
-//"seat_heater_left": 0,
-//"seat_heater_rear_center": 0,
-//"seat_heater_rear_left": 0,
-//"seat_heater_rear_right": 0,
-//"seat_heater_right": 0,
-//"side_mirror_heaters": false,
-//"steering_wheel_heat_level": 0,
-//"steering_wheel_heater": false,
-//"supports_fan_only_cabin_overheat_protection": true,
-//"timestamp": 1692141038419,
-//"wiper_blade_heater": false
