+++ internal/api/models/apiResponse.go
@@ -1,0 +1,13 @@
+package models
+
+import (
+	"encoding/json"
+	"sync"
+)
+
+type ApiResponse struct {
+	Wait     *sync.WaitGroup
+	Result   bool
+	Error    string
+	Response json.RawMessage
+}
