+++ internal/api/models/statesConverter.go
@@ -1,4 +1,4 @@
-package converter
+package models
 
 import (
 	"strings"
