+++ internal/ble/control/key.go
@@ -11,6 +11,7 @@ import (
 	"time"
 
 	"github.com/charmbracelet/log"
+	"github.com/wimaha/TeslaBleHttpProxy/internal/tesla/commands"
 )
 
 func CreatePrivateAndPublicKeyFile() error {
@@ -85,16 +86,16 @@ func RemoveKeyFiles() (error, error) {
 func SendKeysToVehicle(vin string) error {
 	tempBleControl := &BleControl{
 		privateKey:   nil,
-		commandStack: make(chan Command, 1),
+		commandStack: make(chan commands.Command, 1),
 	}
 
 	ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
 	defer cancel()
-	cmd := &Command{
+	cmd := &commands.Command{
 		Command: "add-key-request",
 		Vin:     vin,
 	}
-	conn, car, _, err := tempBleControl.tryConnectToVehicle(ctx, cmd)
+	conn, car, _, err := tempBleControl.TryConnectToVehicle(ctx, cmd)
 	if err == nil {
 		//Successful
 		defer conn.Close()
@@ -102,7 +103,7 @@ func SendKeysToVehicle(vin string) error {
 		defer car.Disconnect()
 		defer log.Debug("disconnect vehicle (A)")
 
-		_, err := tempBleControl.executeCommand(car, cmd)
+		_, err := tempBleControl.ExecuteCommand(car, cmd)
 		if err != nil {
 			return err
 		}
