+++ internal/tesla/commands/commands.go
@@ -1,0 +1,4 @@
+package commands
+
+var ExceptedCommands = []string{"vehicle_data", "auto_conditioning_start", "auto_conditioning_stop", "charge_port_door_open", "charge_port_door_close", "flash_lights", "wake_up", "set_charging_amps", "set_charge_limit", "charge_start", "charge_stop", "session_info"}
+var ExceptedEndpoints = []string{"charge_state", "climate_state"}
